shim: split DecompressedKernel from TritonKernel
