+++ include/aotriton/_internal/triton_kernel.h
@@ -5,18 +5,10 @@
 #define AOTRITON_V2_API_TRITON_KERNEL_H
 
 #include "../runtime.h"
-#include <aotriton/config.h>
+#include "binary_kernel.h"
 #include <memory>
-#include <string_view>
-#include <shared_mutex>
 #include <tuple>
-#include <unordered_map>
 #include <vector>
-
-#include <filesystem>
-
-using pstring_type = std::filesystem::path::string_type;
-using pstring_view = std::basic_string_view<std::filesystem::path::value_type>;
 
 namespace AOTRITON_NS {
 
@@ -65,9 +57,7 @@ struct TritonKernelCompactMeta {
 //                        objects.
 //
 
-class PackedKernel;
-
-class TritonKernel {
+class TritonKernel : public DecompressedKernel {
 public:
   struct Essentials {
     const void* image = nullptr;
@@ -104,38 +94,16 @@ public:
                            size_t sizeof_struct,
                            hipStream_t stream);
 
-  void clear_decompressed_image();
 
 #if AOTRITON_BUILD_FOR_TUNING
   // Will not work unless invoke is called at least once, i.e., If-and-only-iF decompressed
   Essentials get_image_info_iff_decompressed() const;
 #endif
 private:
-  std::tuple<hipFunction_t, hipError_t> load_for_device(int device_id,
-                                                        std::string_view kernel_function_name,
-                                                        std::string_view stem_name,
-                                                        pstring_view package_path);
-  hipFunction_t cfind_function(int device_id) const;
 
   uint64_t blake2b_; // TODO: sanity check of assemblied stem name
   std::string_view ksig_psel_; // psel_name component
   std::string_view ksig_copt_; // copt_name component
-  struct DeviceFunction {
-    DeviceFunction(int device_id_, hipModule_t mod_, hipFunction_t func_);
-    ~DeviceFunction();
-    int device_id = -1;
-    hipModule_t mod = nullptr;
-    hipFunction_t func = nullptr;
-  };
-  std::unordered_map<int, DeviceFunction> funcache_;
-  std::shared_mutex funcache_mutex_;
-
-  Essentials essentials_;
-  bool kernel_loaded_ = false;
-  void decompress_kernel(pstring_view package_path,
-                         std::string_view stem_name);
-  std::shared_ptr<PackedKernel> packed_kernel_ = nullptr;
-  std::shared_mutex packedkernel_mutex_;
 };
 
 struct TritonAuxiliaryArguments {
