+++ v3src/triton_kernel.cc
@@ -1,8 +1,8 @@
 // Copyright © 2023-2025 Advanced Micro Devices, Inc.
 // SPDX-License-Identifier: MIT
 
-#include <aotriton/_internal/packed_kernel.h>
 #include <aotriton/_internal/triton_kernel.h>
+#include <aotriton/_internal/util.h>
 #include <aotriton/runtime.h>
 #include <iostream>
 #include <string>
@@ -20,13 +20,6 @@
 
 #define STRINGIFICATION(s) STRINGIFICATION_I(s)
 #define STRINGIFICATION_I(s) #s
-
-#define AOTRITON_HIP_CHECK_RETURN(expr)                                                                      \
-  do {                                                                                                       \
-    auto r = (expr);                                                                                         \
-    if (r != hipSuccess)                                                                                     \
-      throw std::runtime_error("FAILURE at Line " STRINGIFICATION(__LINE__));                                \
-  } while (0)
 
 namespace AOTRITON_NS {
 
@@ -216,101 +209,11 @@ TritonKernel::direct_invoke(std::string_view mangled_kernel_function_name,
                                reinterpret_cast<void**>(&config));
 }
 
-hipFunction_t
-TritonKernel::cfind_function(int device_id) const {
-  auto iter = funcache_.find(device_id);
-  if (iter == funcache_.end())
-    return nullptr;
-  return iter->second.func;
-}
-
-std::tuple<hipFunction_t, hipError_t>
-TritonKernel::load_for_device(int device_id,
-                              std::string_view kernel_function_name,
-                              std::string_view stem_name,
-                              pstring_view package_path) {
-  hipJitOption opt[] = { hipJitOptionErrorLogBufferSizeBytes,
-                         hipJitOptionErrorLogBuffer,
-                         hipJitOptionInfoLogBufferSizeBytes,
-                         hipJitOptionInfoLogBuffer,
-                         hipJitOptionLogVerbose };
-  const unsigned int errbufsize = 8192;
-  const unsigned int logbufsize = 8192;
-  std::vector<char> err(errbufsize, 0);
-  std::vector<char> log(errbufsize, 0);
-  void* optval[] = { (void*)(uintptr_t)err.size(),
-                     err.data(),
-                     (void*)(uintptr_t)log.size(),
-                     log.data(),
-                     (void*)(uintptr_t)1 };
-
-#if AOTRITON_KERNEL_VERBOSE
-#if defined(_WIN32)
-  std::wcerr << L"Trying to decompress kernel " << package_path;
-  std::cerr << " " << stem_name << std::endl;
-#else
-  std::cerr << "Trying to decompress kernel " << package_path << " " << stem_name << std::endl;
-#endif
-#endif
-  decompress_kernel(package_path, stem_name);
-#if AOTRITON_KERNEL_VERBOSE
-  std::cerr << "Decompress kernel to " << essentials_.image << std::endl;
-#endif
-  if (!essentials_.image) {
-    return std::make_tuple(nullptr, hipErrorInvalidImage);
-  }
-  hipModule_t mod;
-  hipFunction_t func;
-  AOTRITON_HIP_CHECK_RETURN(hipModuleLoadDataEx(&mod, essentials_.image, 5, opt, optval));
-  AOTRITON_HIP_CHECK_RETURN(hipModuleGetFunction(&func, mod, kernel_function_name.data()));
-  funcache_.emplace(std::piecewise_construct,
-                    std::forward_as_tuple(device_id),
-                    std::forward_as_tuple(device_id, mod, func));
-  return std::make_tuple(func, hipSuccess);
-}
-
-void
-TritonKernel::clear_decompressed_image() {
-  std::unique_lock lock(packedkernel_mutex_);
-  essentials_.image = nullptr;
-  packed_kernel_.reset();
-}
-
 #if AOTRITON_BUILD_FOR_TUNING
 TritonKernel::Essentials
 TritonKernel::get_image_info_iff_decompressed() const {
   return essentials_;
 }
 #endif
-
-// kernel_loaded_ is essential. When build for tuning, it is not possible to
-// tell if a kernel is loaded, or the kernel image failed to compile and thus
-// does not exists from beginning by testing essentials_.image == nullptr
-void
-TritonKernel::decompress_kernel(pstring_view package_path,
-                                std::string_view stem_name) {
-  if (kernel_loaded_) {
-    return ;
-  }
-
-  std::unique_lock lock(packedkernel_mutex_);
-  // Check again, another thread may have updated this when this thread is
-  // waiting for the lock
-  if (kernel_loaded_) {
-    return ;
-  }
-  if (!packed_kernel_) {
-    packed_kernel_ = PackedKernel::open(package_path);
-  }
-  if (packed_kernel_) { // open may fail
-#if AOTRITON_KERNEL_VERBOSE
-    std::cerr << "PackedKernel::open returns " << packed_kernel_.get()
-              << " status: " << packed_kernel_->status() << std::endl;
-#endif
-    essentials_ = packed_kernel_->filter(stem_name);
-  }
-  // FIXME: There should be a memory barrier here for non-X86 CPUs.
-  kernel_loaded_ = true;
-}
 
 }
